Fix local import issue
@theMagicalKarp I found out the issue. With Python, when one uses the
import that looks like this: `from base.models.datastore import
DataStore`, it goes through all the __init__.py files of the modules.
Thus it made a circular import because in the __init__ it imports all
the files (including the DataStore class). To fix this, just import
relatively, so `from .datastore import DataStore` will fix it because it
won't go through the __init__.py and thus no circular imports.

Basically the rule is when one has a class that depends on other
classes which depends on the previous class, remove redundancy with the
circular imports by having only local imports in one class. In this
case, the class is DataStore because it imports the models as it needs
it.